backend/requirements.txt
fastapi
uvicorn[standard]
instagrapi
google-generativeai
python-dotenv
supabase
pypdf
websockets
httpx
pydantic
Pillow
python-multipart
